requirements.txt
pdfplumber==0.10.3
sentence-transformers==2.6.0
transformers==4.41.1
torch>=1.13.0
openai>=1.0.0
scikit-learn>=1.3.0
spacy==3.7.4
networkx==3.3
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.2
pydantic>=2.0.0
pandas>=2.0.0
faiss-cpu>=1.7.4
python-docx>=0.8.11
Pillow>=10.0.0
pytesseract>=0.3.10
opencv-python>=4.8.0
openpyxl>=3.1.2
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
reportlab>=4.0.0
